Refactor actionBase.ts
