+++ _common/gameModel/libs/ClientScript/UIfacade/actorFacade.ts
@@ -7,7 +7,6 @@ import { openOverlayItem } from '../gameMap/mapEntities';
 import { getPlayerRolesSelf } from '../multiplayer/multiplayerManager';
 import * as TaskFacade from './taskFacade';
 import { isOngoingAndStartedAction } from '../game/common/simulationState/actionStateAccess';
-import { OnTheRoadAction } from '../game/common/actions/actionBase';
 import {
   getTypedInterfaceState,
   InterfaceState,
@@ -16,6 +15,7 @@ import {
 import { canActorPlanAction } from '../gameInterface/main';
 import { getTranslation } from '../tools/translation';
 import { refreshActivableLayer, refreshSelectionLayer } from '../gameMap/main';
+import { OnTheRoadAction } from '../game/common/actions/actionActors';
 
 /**
  * @returns All currently present actors
