+++ _common/gameModel/libs/ClientScript/game/common/actions/actionTemplateDescriptor/descriptors/mapChoiceTemplate.ts
@@ -3,9 +3,7 @@ import { mainSimLoaderLogger } from '../../../../../tools/logger';
 import { getMapEntityDescriptor } from '../../../../loaders/mapEntitiesLoader';
 import { VehicleType } from '../../../resources/resourceType';
 import { LOCATION_ENUM } from '../../../simulationState/locationState';
-import {
-  SimFlag,
-} from '../../actionTemplate/actionTemplateBase';
+import { SimFlag } from '../../actionTemplate/actionTemplateBase';
 import { ChoiceDescriptor } from '../../choiceDescriptor/choiceDescriptor';
 import { ITemplateDescriptor } from '../templateDescriptor';
 import {
