+++ _common/gameModel/libs/ClientScript/game/common/localEvents/localEventHospital.ts
