+++ _common/gameModel/libs/ClientScript/game/common/localEvents/localEventResourceArrival.ts
